additional fixes

+++ top_earning_players_sql.py
@@ -4,7 +4,7 @@ Created on Mon Feb 9 00:29:26 2020
 @author: Ken Mwaura
 """
 
-from sqlalchemy import Column, String, Numeric, Date, Integer
+from sqlalchemy import Column, String, Numeric, Date, Integer, Float
 
 from base_sql import Base
 
@@ -15,7 +15,7 @@ class Earning_Player(Base):
     id = Column(Integer, primary_key=True)
     username = Column(String(80))
     place = Column(String())
-    earnings = Column(Numeric())
+    earnings = Column(Float())
 
     def __init__(self, username, place, earnings):
         self.place = place
